Modification to allow for placement of Structures
Method created for adding structures. Add the structure to the square
the upper left corner will be located in. Using the height and witch of
the structure the squares contained by the structure are set to Non
Moveable locations. The flooring enum had to be modified to include
Movable and non-mOnable interactions for this; which led to changes in
many of the classes. This use of flooring seems more meaningful
anyhow...